% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_con.R
\name{with_con}
\alias{with_con}
\title{Manage DBI Connections with a 'with' like construct}
\usage{
with_con(driver, code_block, ...)
}
\arguments{
\item{driver}{A DBI driver object, e.g., \code{duckdb::duckdb()}.}

\item{code_block}{A code block to be executed with the created connection.}

\item{...}{Named arguments to be passed to \code{dbConnect} along with the driver.}
}
\value{
The result of the evaluated code block.
}
\description{
The \code{with_con} function creates a database connection using the provided specifications,
passes the connection to the provided code block, and ensures the connection is closed
after the code block is executed.
}
\examples{
if (requireNamespace("duckdb", quietly = TRUE)) {
  library(DBI)
  result <- with_con(duckdb::duckdb(), dbname = ":memory:", {
    dbWriteTable(con, "mtcars", mtcars)
    dbReadTable(con, "mtcars")
  })
  print(result)
}
}
