% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withCon.R
\name{withCon}
\alias{withCon}
\title{Make sure connections are closed}
\usage{
withCon(..., do = { })
}
\arguments{
\item{...}{arbitrary number of connection objects;
should be named to refer to them in the operation}

\item{do}{operations to conduct with the connection objects}
}
\value{
outcome of \code{expr}
}
\description{
This function mimics Python's \code{with} clause, which helps
to make sure that connections are closed.
Concretely, this function closes connection objects when
one of the followings occurred:
\itemize{
\item{operations are completed}
\item{operations throw an exception}
}
}
\details{
Internally, generic function \code{\link{conclose}} is called for
each connection object after operations.
Currently, \code{conclose} is defined for
\code{\link[base]{connection}} and
\code{\link[DBI]{DBIConnection-class}}.
The function still works with unsupported connection object,
for which \code{conclose} does nothing.
}
\examples{
\dontrun{
# database
library(DBI)
withCon(conn = dbConnect(RSQLite::SQLite(), 'temp.db'), do = {
  dbWriteTable(conn, 'tbl', mtcars)
  dbReadTable(conn, 'tbl')
})
file.remove('temp.db')

# files
withCon(f1 = file('foo.txt', 'w'), f2 = file('bar.txt', 'w'), do = {
  write('foo', f1)
  write('bar', f2)
})
file.remove('foo.txt', 'bar.txt')
}
}
