% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withCon.R
\name{conclose}
\alias{conclose}
\alias{conclose.connection}
\alias{conclose.DBIConnection}
\title{Generic `close` operation for connection objects}
\usage{
conclose(obj)

\method{conclose}{connection}(obj)

\method{conclose}{DBIConnection}(obj)
}
\arguments{
\item{obj}{connection object}
}
\description{
Generic `close` operation for connection objects
}
\examples{
f <- file('foo.txt', 'w')
conclose(f)  # same as `close(f)`

g <- DBI::dbConnect(RSQLite::SQLite(), "bar.db")
conclose(g)  # same as `DBI::dbDisconnect(g)`
}
