% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withCon.R
\name{withCon}
\alias{withCon}
\title{Make sure connection is closed}
\usage{
withCon(..., do = { })
}
\arguments{
\item{...}{arbitrary number of connection objects;
should be named to refer to them in the operation;
currently \code{\link[base]{connection}} and
\code{\link[DBI]{DBIConnection-class}} are supported}

\item{do}{operations to conduct with the connection objects}
}
\value{
outcome of \code{expr}
}
\description{
This function mimics Python's \code{with} clause.  Concretely,
Connections objects are automatically closed:
\itemize{
\item{After a series of operations are conducted}
\item{Operations throw exceptions}
}
}
\examples{
\dontrun{
# database
library(DBI)
withCon(conn=dbConnect(RSQLite::SQLite(), 'temp.db'), do = {
  dbWriteTable(conn, 'tbl', mtcars)
  dbReadTable(conn, 'tbl')
})
file.remove('temp.db')

# files
withCon(f1 = file('foo.txt', 'w'), f2 = file('bar.txt', 'w'), do = {
  write('foo', f1)
  write('bar', f2)
})
file.remove('foo.txt', 'bar.txt')
}
}
